+++ packages/bloc_test/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 7.0.0-dev.2
+
+- **BREAKING**: `blocTest` make `build` synchronous
+- fix: `blocTest` improve `wait` behavior when debouncing, etc...
+- feat: `blocTest` do not require `async` on `act` and `verify`
+
 # 7.0.0-dev.1
 
 - **BREAKING**: upgrade to `bloc ^6.0.0-dev.1`